* Development 12 | Beta 1 *
* Updated to Version 1.0.1.8
* Renamed TextDisplayArea to GUITextDisplayArea
* Fixed bug: getPercentageComplete() in GUILoadingBar only giving the
values 0.0 and 100.0
* Fixed bug: Plain Java code didn't work unless the LWJGL library
was present
* Added methods in Timer to improve it:
  - stop()
  - reset()
* ANDROID SUPPORT!!!!!!!! (Including GUI :) !!!!!!!)
* Added android support for the Image class
* Added android support to the BasicRenderer
* Added BasicRendererAndroid
* Added AndroidDisplay
* Added AndroidGame
* Added AndroidGameThread
* Added AndroidStore
* Added AndroidActivity
* Added AndroidSettings
* Added ScreenUtils
* Added AndroidInput
* Added AndroidGestureListener
* Added AndoidTouchListener
* Added AndroidSensor
* Added the method println(String message) in Logger
* Added getAndroidBounds() method in Entity2D
* Added the colour orange
* The logger now logs differently when using android (Log.d(String tag ,
String msg))
* Found an issue: If you set a variable to a colour constant then change
its a,r,g,b value the colour constant will change with the variable
* From now on when downloading the engine you will need to download
extension jars to allow the use of Android or OpenGL

+++ MyEngine-Master/src/org/simplecorporation/myengine/core/gui/loadingbar/GUILoadingBar.java
@@ -53,7 +53,7 @@ public abstract class GUILoadingBar extends GUIComponent {
 	/* Returns the percentage of the process that is completed */
 	public double getPercentageComplete() {
 		//Return the percentage
-		return (this.currentLoadingStage / this.loadingStages) * 100;
+		return (double)((double)this.currentLoadingStage / (double)this.loadingStages) * 100;
 	}
 	
 }
